+++ modules/juce_core/containers/juce_HashMap.h
@@ -43,7 +43,7 @@ public:
     /** Generates a simple hash from an integer. */
     static int generateHash (const int key, const int upperLimit) noexcept        { return std::abs (key) % upperLimit; }
     /** Generates a simple hash from a string. */
-    static int generateHash (const String& key, const int upperLimit) noexcept    { return (int) (((uint32) key.hashCode()) % upperLimit); }
+    static int generateHash (const String& key, const int upperLimit) noexcept    { return (int) (((uint32) key.hashCode()) % (uint32) upperLimit); }
     /** Generates a simple hash from a variant. */
     static int generateHash (const var& key, const int upperLimit) noexcept       { return generateHash (key.toString(), upperLimit); }
 };
