+++ modules/juce_core/xml/juce_XmlElement.h
@@ -605,7 +605,7 @@ public:
 
         if (num > 1)
         {
-            HeapBlock <XmlElement*> elems (num);
+            HeapBlock <XmlElement*> elems ((size_t) num);
             getChildElementsAsArray (elems);
             sortArray (comparator, (XmlElement**) elems, 0, num - 1, retainOrderOfEquivalentItems);
             reorderChildElements (elems, num);
